🐓📕 Checkpoint ./server.js:26780611/1551 ./commands/Misc/afk.js:26780611/137

+++ server.js
@@ -238,26 +238,21 @@ MANAGE_EMOJIS'*/
 
     client.logger.error(error);
   }
+   /*====================================================================*/
+  //<AFK CMD>
   if(db.has(`afk-${message.author.id}+${message.guild.id}`)) {
-
         const info = db.get(`afk-${message.author.id}+${message.guild.id}`)
-
+       const mes = args.join(" ")
+       if(!mes){
         await db.delete(`afk-${message.author.id}+${message.guild.id}`)
-
-        message.reply(`Your afk status have been removed (${info})`)
-
-    }
-
+        
+    message.reply(`Your afk status have been removed (${info})`)
+    }}
     //checking for mentions
-
     if(message.mentions.members.first()) {
-
         if(db.has(`afk-${message.mentions.members.first().id}+${message.guild.id}`)) {
-
             message.channel.send(message.mentions.members.first().user.tag + ":" + db.get(`afk-${message.mentions.members.first().id}+${message.guild.id}`))
-
         }else return;
-
     }else;
   /*====================================================================*/
   //<COMMAND EP/LEVEL>
